Makefile Update

+++ ft_atol.c
@@ -6,7 +6,7 @@
 /*   By: anarodri <anarodri@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/28 14:02:10 by anarodri          #+#    #+#             */
-/*   Updated: 2022/05/19 14:29:11 by anarodri         ###   ########.fr       */
+/*   Updated: 2022/05/25 12:23:42 by anarodri         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 long int	ft_atol(const char *str)
 {
-	long long	nb;
-	int			i;
-	int			sign;
+	long nb;
+	int		i;
+	int		sign;
 
 	nb = 0;
 	i = 0;
